update OSLogger impl
