+++ Tests/ShopifyCheckoutSheetKitTests/Mocks/MockLogger.swift
@@ -21,10 +21,9 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-import os.log
 @testable import ShopifyCheckoutSheetKit
 
-class MockLogger: ProductionLogger {
+class MockLogger: NoOpLogger {
 	var loggedError: Error?
 	var loggedMessage: String?
